+++ CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Send Empty parameter instead of Node in the detach rpc.
 - Make the `grpc_socket_port` field optional for the `ConfigMessage`, the default port will be
   `50051` on the server.
+- The Attach rpc now returns a `MessageHubEvent`, which can either be an error or an Astarte message.
 
 ## [0.6.2] - 2024-04-23